integrate server-side gash into `.shtml` pages
Fixes #2

This is an ugly, ugly bit of code that parses the file byte-by-byte
looking for gash commands before writing it to the stream. If it found a
command, then it runs it at the gash shell and writes the output to the
stream.

On the wacky logic here:

It's ugly to look at, and I know that. I tried to alleviate this with
comments, but they only help so much.

It's done on a single-byte basis for a couple of reasons: this allows
incremental reading of the `.shtml` file and writing it to the stream,
so the stream is written to as the file is read (even though we might
find gash commands along the way!).

The other reason it is done as a single-byte iteration is because I want
to know if this offers any performance improvements over reading
multiple bytes at a time and checking if the beginning of a comment is
contained within those bytes. I might not get around to rewriting it
this way and benchmarking it, but it's a potential room for performance
comparison/boost anyway.

The problem with reading multiple bytes is that it seems to require
almost the same code smell logic - every time you find something that
looks like the start of a gash command, you have to deal with reading
the *next* bytes too and checking if those could be a gash command (in
case the gash comment is only partially contained in the bytes you
found, and you need to read more to be sure it's actually one of them).
You don't know if the perceived comment is *actually* a gash command
until the very end of the comment is seen, so the actual benefit of
doing it that way is questionable if you really want to write to the
stream as the file is read.

In spite of the above, this commit is probably the wrong approach -
right now, the code is doing a bunch of branch conditionals for every
byte it encounters. If we buffered, say, 32 bytes at a time and just ran
the comment checks against those bytes, it would likely be faster since
branches are slow and we could cut the number of branch conditions
executed by perhaps a factor of 32. The actual decrease in branches
executed of course depends on the way Rust does its equality tests on
byte arrays, and the means by which we iterate over those byte arrays to
look for chunks that could be part of an `<!-- #exec cmd`.

Future:

A possible improvement to this logic is to allow any number of spaces in
the comment block itself. Right now, it's only expecting exactly one
space, and if you have any more than that it won't register as a gash
command and will simply write the `.shtml` comment block itself to the
stream instead of running the command.