Add static file streaming and caching
Static files are now served by default in a streaming manner, and are
cached the first time they are read. During this first instance, the
file is read 4096 bytes at a time and written to the TcpStream as it is
read.

After the first time, the file is cached and is stored in a HashMap
inside a MutexArc. 4096 bytes are read from the MutexArc before sending
them to the client. The byte value here and above are easy enough to
tweak.

On my machine, this starts producing errors around ~5000 concurrent
connections requesting a randomly generated 5 KiB binary file. That's
not a good number, but it's considerably better than the 3000 or so
that the default implementation will handle.

An important distinction between this and the starting code is that
this will not "hard crash" when over the limit. The starting code
deadlocks once I start benchmarking with a rate of ~3000 on a 5KiB
file and will never recover. This will recover, and new files may be
served once the rate drops.